+++ NetIdMobileSdk/NetIdMobileSdk/api/NetIdConfig.swift
@@ -14,18 +14,48 @@
 
 import Foundation
 
+public struct LoginLayerConfig {
+    public var headlineText: String
+    public var loginText: String
+    public var continueText: String
+    
+    public init () {
+        self.headlineText = ""
+        self.loginText = ""
+        self.continueText = ""
+    }
+}
+
+public struct PermissionLayerConfig {
+    public var logoId: String
+    public var headlineText: String
+    public var legalText: String
+    public var continueText: String
+    
+    public init() {
+        self.logoId = ""
+        self.headlineText = ""
+        self.legalText = ""
+        self.continueText = ""
+    }
+}
+
 public struct NetIdConfig {
     public var host: String
     public var clientId: String
     public var redirectUri: String
     public var originUrlScheme: String
     public var claims: [String: String]?
+    public var loginLayerConfig: LoginLayerConfig?
+    public var permissionLayerConfig: PermissionLayerConfig?
 
-    public init(host: String, clientId: String, redirectUri: String, originUrlScheme: String, claims: [String: String]?) {
+    public init(host: String, clientId: String, redirectUri: String, originUrlScheme: String, claims: [String: String]?, loginLayerConfig: LoginLayerConfig?, permissionLayerConfig: PermissionLayerConfig?) {
         self.host = host
         self.clientId = clientId
         self.redirectUri = redirectUri
         self.originUrlScheme = originUrlScheme
         self.claims = claims
+        self.loginLayerConfig = loginLayerConfig
+        self.permissionLayerConfig = permissionLayerConfig
     }
 }