+++ NetIdMobileSdk/NetIdMobileSdk/api/NetIdService.swift
@@ -100,11 +100,13 @@ open class NetIdService: NSObject {
         }
         switch authFlow {
         case .Permission:
+            let config = netIdConfig?.permissionLayerConfig
             return AnyView(AuthorizationSoftView(delegate: self, presentingViewController: currentViewController,
-                    appIdentifiers: netIdApps))
+                                                 appIdentifiers: netIdApps, logoId: (config?.logoId) ?? "", headlineText: (config?.headlineText) ?? "", legalText: (config?.legalText) ?? "", continueText: (config?.continueText) ?? ""))
         case .LoginPermission, .Login:
+            let config = netIdConfig?.loginLayerConfig
             return AnyView(AuthorizationHardView(delegate: self, presentingViewController: currentViewController,
-                    appIdentifiers: netIdApps))
+                                                 appIdentifiers: netIdApps, headlineText: (config?.headlineText) ?? "", loginText: (config?.loginText) ?? "", continueText: (config?.continueText) ?? ""))
         }
 //        case .Soft:
 //            return AnyView(AuthorizationSoftView(delegate: self, presentingViewController: currentViewController,